+++ Magic/CHANGELOG.md
@@ -19,6 +19,7 @@
    - Automata are reflective, except for the redstone
    - Re-balance automata drops and resurrection costs to avoid abuse
    - Removed automata.yml, automata now just despawn on chunk onload and server restart
+   - Defenders use a different heart (lapis block)
  - Blast and Collapse are now projectiles
  - SP Earn Bonus (Mana Talisman) will work on SP item pickups
  - Mana Talisman set to keep on death